typeahead: clean-up
